Add SSH Config section to tpcds-env.sh
The purpose of this section is to allow the user to configure whether they need ssh keys for access between nodes.  Currently this is not an option in the existing code, but this config will give users the ability to enable ssh keys to be used (by making SSH_KEY=True) and then provide the name of the shared ssh key that is set for the selected $USER across all of the data nodes.